Exceptions handling

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.example.demo.student;
 
+import com.example.demo.student.exception.BadRequestException;
+import com.example.demo.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,12 +18,20 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        //TODO validation if email is already taken
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " is taken");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
-        //TODO check if student exists
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id + " + studentId + " does not exits");
+        }
         studentRepository.deleteById(studentId);
     }
 }
